refactor: split api.ts into client, errors and schoolName modules
