+++ src/services/api.ts
@@ -1,664 +1,32 @@
 /**
  * API Service for PyrePortal
  * Handles all communication with the backend API
+ *
+ * Split into modules: error mapping lives in ./apiErrors, the HTTP client and
+ * auth headers in ./apiClient, school-name state in ./schoolName. This module
+ * keeps the endpoint methods and types and re-exports the public surface so
+ * existing imports keep working.
  */
-
-import { adapter } from '@platform';
 
 import { createLogger } from '../utils/logger';
 
+import { apiCall, buildAuthHeaders, getApiBaseUrl, hasDeviceApiKey } from './apiClient';
+import { isNotFoundError, mapAttendanceErrorToGerman, mapServerErrorToGerman } from './apiErrors';
+
+export {
+  ApiError,
+  formatRoomName,
+  isNetworkRelatedError,
+  isNotFoundError,
+  isWCRoomAlias,
+  mapApiErrorToGerman,
+  mapServerErrorToGerman,
+  WC_ROOM_ALIASES,
+} from './apiErrors';
+export { initializeApi, setNetworkStatusCallback } from './apiClient';
+export { fetchSchoolName, getSchoolName, onSchoolNameLoaded } from './schoolName';
+
 const logger = createLogger('API');
-
-/**
- * Structured API error response with optional details
- * Used for rich error responses like capacity exceeded
- */
-interface ApiErrorResponse {
-  status: string;
-  message: string;
-  code?: string;
-  details?: {
-    // Room capacity fields
-    room_id?: number;
-    room_name?: string;
-    current_occupancy?: number;
-    max_capacity?: number;
-    // Activity capacity fields
-    activity_id?: number;
-    activity_name?: string;
-    current_participants?: number;
-    max_participants?: number;
-    // Duplicate-active-visit fields (Issue #844, backend STUDENT_ALREADY_ACTIVE)
-    student_id?: number;
-    existing_visit_id?: number;
-    entry_time?: string;
-  };
-}
-
-/**
- * Custom error class that preserves structured API error data
- */
-export class ApiError extends Error {
-  public readonly code?: string;
-  public readonly details?: ApiErrorResponse['details'];
-  public readonly statusCode: number;
-
-  constructor(
-    message: string,
-    statusCode: number,
-    code?: string,
-    details?: ApiErrorResponse['details']
-  ) {
-    super(message);
-    this.name = 'ApiError';
-    this.statusCode = statusCode;
-    this.code = code;
-    this.details = details;
-  }
-}
-
-/**
- * Error message mapping entry.
- * Pattern can be a single string or array of strings (OR logic).
- */
-type ErrorMapping = readonly [pattern: string | readonly string[], germanMessage: string];
-
-/**
- * Error message mappings ordered by specificity.
- *
- * IMPORTANT: Order matters! Specific messages MUST come before generic patterns.
- * The first matching pattern wins.
- *
- * Backend error messages sourced from:
- * - /backend/auth/device/errors.go
- * - /backend/api/iot/common/errors.go
- * - /backend/api/iot/checkin/workflow.go
- * - /backend/api/iot/sessions/handlers.go
- * - /backend/api/iot/attendance/handlers.go
- * - /backend/api/iot/rfid/handlers.go
- * - /backend/api/iot/feedback/handlers.go
- */
-const ERROR_MESSAGE_MAPPINGS: readonly ErrorMapping[] = [
-  // 1. CAPACITY ERRORS (409)
-  // Backend sends both lowercase (Go errors) and capitalized (JSON Message field)
-  [
-    ['ACTIVITY_CAPACITY_EXCEEDED', 'activity capacity exceeded', 'Activity capacity exceeded'],
-    'Aktivität ist voll. Maximale Teilnehmerzahl erreicht.',
-  ],
-  [
-    ['ROOM_CAPACITY_EXCEEDED', 'room capacity exceeded', 'Room capacity exceeded'],
-    'Raum ist voll. Kein Platz mehr verfügbar.',
-  ],
-  // Duplicate active visit (Issue #844, backend migration 1.15.45 + checkin/workflow.go)
-  [
-    ['STUDENT_ALREADY_ACTIVE', 'student already has an active visit'],
-    'Schüler*in ist bereits angemeldet.',
-  ],
-
-  // 2. AUTHENTICATION ERRORS (401)
-  ['invalid device API key', 'API-Schlüssel ungültig. Bitte Geräte-Konfiguration prüfen.'],
-  ['device API key is required', 'API-Schlüssel nicht konfiguriert. Bitte .env Datei prüfen.'],
-  ['invalid API key format', 'API-Schlüssel Format ungültig. Bearer Token erwartet.'],
-  ['invalid staff PIN', 'Ungültiger PIN. Bitte erneut versuchen.'],
-  ['staff PIN is required', 'PIN nicht angegeben.'],
-  [
-    'staff account is locked due to failed PIN attempts',
-    'Konto gesperrt wegen zu vieler Fehlversuche. Bitte Administrator kontaktieren.',
-  ],
-  ['maximum PIN attempts exceeded', 'Maximale PIN-Versuche überschritten. Konto gesperrt.'],
-  ['locked', 'Konto gesperrt. Bitte später erneut versuchen.'], // Generic fallback for lock-related
-
-  // 3. AUTHORIZATION ERRORS (403)
-  ['device is not active', 'Gerät ist deaktiviert. Bitte Administrator kontaktieren.'],
-  ['device is offline', 'Gerät ist als offline markiert. Bitte Administrator kontaktieren.'],
-
-  // 4. SESSION ERRORS (400/409)
-  [
-    'device is already running an activity session',
-    'Gerät führt bereits eine Aktivität durch. Bitte zuerst beenden.',
-  ],
-  ['no active session to end', 'Keine aktive Sitzung zum Beenden vorhanden.'],
-  ['no active session', 'Keine aktive Sitzung. Bitte zuerst eine Aktivität starten.'],
-  ['invalid session ID', 'Ungültige Sitzungs-ID.'],
-  ['activity_id is required', 'Aktivität muss ausgewählt werden.'],
-  ['at least one supervisor', 'Mindestens ein Betreuer muss ausgewählt werden.'], // Matches both variants
-
-  // 5. RFID ERRORS (400/404) - Only cases where "Armband nicht zugewiesen" is appropriate
-  [
-    ['RFID tag not found', 'RFID tag not assigned'],
-    'Armband ist nicht zugewiesen. Bitte an Betreuer wenden.',
-  ],
-  [
-    'staff RFID authentication must be done via session management',
-    'Betreuer-Armband kann hier nicht verwendet werden.',
-  ],
-  ['student RFID tag required for pickup query', 'Bitte Schüler-Armband scannen.'],
-  ['RFID parameter is required', 'RFID-Tag fehlt in der Anfrage.'],
-
-  // 6. ATTENDANCE/VISIT ERRORS (404)
-  ['no active visit found for student', 'Kein aktiver Besuch für diesen Schüler gefunden.'],
-  ['person is not a student', 'Person ist kein Schüler.'],
-  ['no active groups in specified room', 'Keine aktiven Gruppen im ausgewählten Raum.'],
-
-  // 7. STAFF ERRORS (400/404)
-  ['invalid staff ID', 'Ungültige Mitarbeiter-ID.'],
-  ['staff not found', 'Mitarbeiter nicht gefunden.'],
-  ['staff has no RFID tag assigned', 'Mitarbeiter hat kein Armband zugewiesen.'],
-
-  // 8. FEEDBACK ERRORS (400/404)
-  ['student_id is required', 'Schüler-ID fehlt.'],
-  ['value is required', 'Bewertung fehlt.'],
-  ['student not found', 'Schüler nicht gefunden.'],
-
-  // 9. VALIDATION ERRORS (400)
-  ['room_id is required for check-in', 'Raum muss für Check-in ausgewählt werden.'],
-  ['tagId parameter is required', 'Tag-ID fehlt in der Anfrage.'],
-  ["destination must be 'zuhause' or 'unterwegs'", "Ziel muss 'zuhause' oder 'unterwegs' sein."],
-  [
-    'destination is required for confirm_daily_checkout',
-    'Ziel muss für endgültiges Auschecken angegeben werden.',
-  ],
-
-  // 10. INTERNAL SERVER ERRORS (500)
-  [
-    'schulhof activity not configured',
-    'Schulhof-Aktivität nicht konfiguriert. Bitte Administrator kontaktieren.',
-  ],
-  [
-    'WC activity not configured',
-    'Toilette-Aktivität nicht konfiguriert. Bitte Administrator kontaktieren.',
-  ],
-  [
-    'WC activity auto-create requires staff context',
-    'Toilette-Aktivität konnte nicht erstellt werden. Bitte zuerst Betreuer-RFID scannen.',
-  ],
-  [
-    'failed to create Schulhof session',
-    'Schulhof-Sitzung konnte nicht erstellt werden. Bitte erneut versuchen.',
-  ],
-  [
-    'failed to create WC session',
-    'Toilette-Sitzung konnte nicht erstellt werden. Bitte erneut versuchen.',
-  ],
-  // Generic fallback — MUST stay AFTER specific 'failed to create Toilette/Schulhof session' entries (substring match)
-  ['failed to create session', 'Sitzung konnte nicht erstellt werden. Bitte erneut versuchen.'],
-  ['failed to create visit record', 'Besuch konnte nicht erstellt werden. Bitte erneut versuchen.'],
-  ['failed to end visit record', 'Besuch konnte nicht beendet werden. Bitte erneut versuchen.'],
-  ['failed to get room information', 'Rauminformationen konnten nicht abgerufen werden.'],
-  ['failed to check room capacity', 'Raumkapazität konnte nicht geprüft werden.'],
-  ['failed to get activity information', 'Aktivitätsinformationen konnten nicht abgerufen werden.'],
-  ['failed to check activity capacity', 'Aktivitätskapazität konnte nicht geprüft werden.'],
-  ['error finding active groups in room', 'Aktive Gruppen im Raum konnten nicht gefunden werden.'],
-  [
-    'failed to get person data for staff',
-    'Personendaten für Mitarbeiter konnten nicht abgerufen werden.',
-  ],
-
-  // 11. GENERIC HTTP STATUS CODES - Fallbacks (must be after specific messages)
-  [['401', 'Unauthorized'], 'Authentifizierung fehlgeschlagen. Bitte erneut anmelden.'],
-  [['403', 'Forbidden'], 'Keine Berechtigung für diese Aktion.'],
-  [['404', 'not found', 'Not Found'], 'Ressource nicht gefunden. Bitte Konfiguration prüfen.'],
-  [['409', 'Conflict'], 'Konflikt bei der Anfrage. Bitte erneut versuchen.'],
-  [['400', 'Bad Request'], 'Ungültige Anfrage. Bitte Eingaben prüfen.'],
-  [
-    ['500', '502', '503', '504', 'Internal Server Error', 'Bad Gateway', 'Service Unavailable'],
-    'Server nicht erreichbar. Bitte später versuchen.',
-  ],
-] as const;
-
-/**
- * Map server error messages to German user-friendly messages.
- * Uses a data-driven approach with ordered pattern matching.
- */
-export function mapServerErrorToGerman(errorMessage: string): string {
-  for (const [patterns, germanMessage] of ERROR_MESSAGE_MAPPINGS) {
-    const patternList = typeof patterns === 'string' ? [patterns] : patterns;
-    if (patternList.some(pattern => errorMessage.includes(pattern))) {
-      return germanMessage;
-    }
-  }
-  return errorMessage; // Fallback - return original for unknown errors
-}
-
-/** Type guard to check if value is a string or number */
-function isStringOrNumber(value: unknown): value is string | number {
-  return typeof value === 'string' || typeof value === 'number';
-}
-
-/**
- * Canonical names a backend toilet special-room can come back with, in
- * canonical-first order. The first entry is what Phoenix auto-creates;
- * the second is an accepted alias a tenant may have created manually.
- *
- * Must stay in sync with Phoenix backend/constants/activities.go
- * (WCRoomName, WCRoomAliasName) and frontend/src/lib/room-helpers.ts
- * (SYSTEM_ROOM_NAMES). Matching is exact-case to mirror the backend's
- * `IsWCRoomName` check. Adding a new alias requires updating all three.
- */
-export const WC_ROOM_ALIASES = ['WC', 'Toilette'] as const;
-
-/**
- * Returns true when the given room name is one of the toilet aliases.
- * Only used internally; exported as a test seam for api.test.ts.
- */
-export function isWCRoomAlias(name: string): boolean {
-  return (WC_ROOM_ALIASES as readonly string[]).includes(name);
-}
-
-/**
- * Map backend room names to German display names.
- * Any toilet-room alias is shown as "Toilette" in the kiosk UI.
- */
-export function formatRoomName(name: string): string {
-  if (isWCRoomAlias(name)) return 'Toilette';
-  return name;
-}
-
-/**
- * Format activity capacity error message from details
- */
-function formatActivityCapacityError(details: Record<string, unknown>): string {
-  const activityName = details.activity_name;
-  const currentParticipants = details.current_participants;
-  const maxParticipants = details.max_participants;
-
-  if (isStringOrNumber(activityName) && isStringOrNumber(maxParticipants)) {
-    const current = isStringOrNumber(currentParticipants) ? currentParticipants : maxParticipants;
-    return `${activityName} ist voll (${current}/${maxParticipants} Teilnehmer).`;
-  }
-  return 'Aktivität ist voll. Maximale Teilnehmerzahl erreicht.';
-}
-
-/**
- * Format room capacity error message from details
- */
-function formatRoomCapacityError(details: Record<string, unknown>): string {
-  const roomName = details.room_name;
-  const currentOccupancy = details.current_occupancy;
-  const maxCapacity = details.max_capacity;
-
-  if (isStringOrNumber(roomName) && isStringOrNumber(maxCapacity)) {
-    const current = isStringOrNumber(currentOccupancy) ? currentOccupancy : maxCapacity;
-    const displayName = typeof roomName === 'string' ? formatRoomName(roomName) : roomName;
-    return `${displayName} ist voll (${current}/${maxCapacity} Plätze belegt).`;
-  }
-  return 'Raum ist voll. Kein Platz mehr verfügbar.';
-}
-
-/**
- * Format duplicate-active-visit error message from details
- * Backend (Issue #844) returns the existing visit's room so the kiosk can
- * tell the user which room the student is already checked into rather than
- * the generic "bereits angemeldet". `room_name` may be missing when the
- * backend's best-effort lookup couldn't resolve the existing visit (rare
- * race window between INSERT failure and response build) — fall back to
- * the generic message in that case.
- */
-function formatStudentAlreadyActiveError(details: Record<string, unknown>): string {
-  const roomName = details.room_name;
-  if (typeof roomName === 'string' && roomName.length > 0) {
-    return `Schüler*in ist bereits angemeldet in ${formatRoomName(roomName)}.`;
-  }
-  return 'Schüler*in ist bereits angemeldet.';
-}
-
-/**
- * Map API errors to German user-friendly messages with rich details support
- * Handles structured error responses (e.g., capacity errors with room/activity details)
- */
-export function mapApiErrorToGerman(error: unknown): string {
-  // Handle non-ApiError cases first
-  if (!(error instanceof ApiError)) {
-    const message = error instanceof Error ? error.message : String(error);
-    return mapServerErrorToGerman(message);
-  }
-
-  // Activity capacity exceeded
-  if (error.code === 'ACTIVITY_CAPACITY_EXCEEDED' && error.details) {
-    return formatActivityCapacityError(error.details);
-  }
-
-  // Room capacity exceeded
-  if (error.code === 'ROOM_CAPACITY_EXCEEDED' && error.details) {
-    return formatRoomCapacityError(error.details);
-  }
-
-  // Duplicate active visit (Issue #844)
-  if (error.code === 'STUDENT_ALREADY_ACTIVE' && error.details) {
-    return formatStudentAlreadyActiveError(error.details);
-  }
-
-  // Fall back to message-based mapping
-  return mapServerErrorToGerman(error.message);
-}
-
-/**
- * Patterns that indicate a network-related error
- * Includes both English (technical) and German (translated) patterns
- */
-const NETWORK_ERROR_PATTERNS = [
-  'network',
-  'netzwerk',
-  'failed to fetch',
-  'fetch',
-  'networkerror',
-  'timeout',
-  'timed out',
-  'connection',
-  'verbindung',
-  'offline',
-];
-
-/**
- * Check if an error (object or message) indicates a network-level error
- * Checks navigator.onLine first, then pattern-matches the error message
- */
-export function isNetworkRelatedError(error: unknown): boolean {
-  if (typeof navigator !== 'undefined' && !navigator.onLine) return true;
-  const message =
-    error instanceof Error ? error.message.toLowerCase() : String(error).toLowerCase();
-  return NETWORK_ERROR_PATTERNS.some(pattern => message.includes(pattern));
-}
-
-/**
- * Map attendance-specific errors to German user-friendly messages
- * Provides context-aware error messages for attendance operations
- *
- * This function first tries to match specific backend messages via
- * mapServerErrorToGerman, then falls back to context-specific generic messages.
- */
-function mapAttendanceErrorToGerman(errorMessage: string, context: 'toggle' | 'feedback'): string {
-  // Network errors - use consolidated handler
-  if (isNetworkRelatedError(errorMessage)) {
-    return 'Netzwerkfehler. Bitte Verbindung prüfen.';
-  }
-
-  // ============================================================
-  // Try specific backend messages first via main mapper
-  // If it returns the original message, it means no specific match was found
-  // ============================================================
-  const specificMessage = mapServerErrorToGerman(errorMessage);
-
-  // Check if a specific mapping was found (not returned unchanged)
-  // We check both exact match and common generic fallbacks
-  const isGenericFallback =
-    specificMessage === errorMessage ||
-    specificMessage === 'Ressource nicht gefunden. Bitte Konfiguration prüfen.' ||
-    specificMessage === 'Keine Berechtigung für diese Aktion.' ||
-    specificMessage === 'Authentifizierung fehlgeschlagen. Bitte erneut anmelden.' ||
-    specificMessage === 'Ungültige Anfrage. Bitte Eingaben prüfen.';
-
-  // If a specific message was found, use it
-  if (!isGenericFallback) {
-    return specificMessage;
-  }
-
-  // ============================================================
-  // Context-specific fallbacks for generic HTTP status codes
-  // ============================================================
-
-  // 404 errors - context-specific messages
-  // Include lowercase 'not found' for consistency with main mapper
-  if (
-    errorMessage.includes('404') ||
-    errorMessage.includes('not found') ||
-    errorMessage.includes('Not Found')
-  ) {
-    switch (context) {
-      case 'toggle':
-        return 'Schüler nicht gefunden. RFID-Tag möglicherweise nicht zugewiesen.';
-      case 'feedback':
-        return 'Feedback-Service nicht erreichbar. Bitte später versuchen.';
-    }
-  }
-
-  // 403 errors - permission denied
-  if (errorMessage.includes('403') || errorMessage.includes('Forbidden')) {
-    switch (context) {
-      case 'toggle':
-        return 'Keine Berechtigung für An-/Abmeldung dieses Schülers.';
-      case 'feedback':
-        return 'Keine Berechtigung für Feedback-Übermittlung.';
-    }
-  }
-
-  // 401 errors - authentication issues
-  if (errorMessage.includes('401') || errorMessage.includes('Unauthorized')) {
-    return 'Authentifizierung fehlgeschlagen. Bitte erneut anmelden.';
-  }
-
-  // 400 errors - bad request
-  if (errorMessage.includes('400') || errorMessage.includes('Bad Request')) {
-    return 'Ungültige Anfrage. Bitte Eingaben prüfen.';
-  }
-
-  // Fallback - use generic mapper result (handles 5xx errors and other cases)
-  return specificMessage;
-}
-
-/**
- * Detect 404 responses. Prefer the structured `ApiError.statusCode` set by
- * `apiCall`. Keep the substring fallback so errors that are not `ApiError`
- * instances (or wrapped messages) still resolve the same way as before.
- */
-export const isNotFoundError = (error: unknown, errorMessage: string): boolean => {
-  if (error instanceof ApiError && error.statusCode === 404) {
-    return true;
-  }
-  return errorMessage.includes('404');
-};
-
-// Environment configuration - will be loaded at runtime
-let API_BASE_URL = '';
-let DEVICE_API_KEY = '';
-
-const schoolNameState = { value: null as string | null };
-let isInitialized = false;
-
-// Network status callback - set by the app to receive status updates from API calls
-type NetworkStatusCallback = (quality: 'online' | 'poor' | 'offline', responseTime: number) => void;
-let networkStatusCallback: NetworkStatusCallback | null = null;
-
-const POOR_THRESHOLD_MS = 1000;
-
-/**
- * Register a callback to receive network status updates from API calls
- */
-export function setNetworkStatusCallback(callback: NetworkStatusCallback | null): void {
-  networkStatusCallback = callback;
-}
-
-/**
- * Report network status based on API call result
- */
-function reportNetworkStatus(responseTime: number, success: boolean): void {
-  if (!networkStatusCallback) {
-    logger.debug('Network status callback not registered, skipping update');
-    return;
-  }
-
-  let quality: 'online' | 'poor' | 'offline';
-  if (success) {
-    quality = responseTime > POOR_THRESHOLD_MS ? 'poor' : 'online';
-  } else {
-    quality = 'offline';
-  }
-  logger.debug('Reporting network status', { quality, responseTime, success });
-
-  networkStatusCallback(quality, responseTime);
-}
-
-/**
- * Initialize API configuration via platform adapter
- */
-export async function initializeApi(): Promise<void> {
-  if (isInitialized) {
-    return;
-  }
-
-  await adapter.loadConfig();
-  API_BASE_URL = adapter.getApiBaseUrl();
-  DEVICE_API_KEY = adapter.getDeviceApiKey();
-  isInitialized = true;
-
-  logger.info('API initialized with platform configuration', {
-    platform: adapter.platform,
-    baseUrl: API_BASE_URL,
-    hasApiKey: !!DEVICE_API_KEY,
-  });
-}
-
-/**
- * Returns the school name for the device, or null if not yet loaded / unavailable.
- */
-export function getSchoolName(): string | null {
-  return schoolNameState.value;
-}
-
-/**
- * Register a listener that fires once when the school name becomes available.
- * If the name is already loaded, the listener is called synchronously.
- * Returns an unsubscribe function.
- */
-export function onSchoolNameLoaded(listener: (name: string) => void): () => void {
-  if (schoolNameState.value) {
-    listener(schoolNameState.value);
-    return () => undefined;
-  }
-  schoolNameListeners.add(listener);
-  return () => schoolNameListeners.delete(listener);
-}
-
-const schoolNameListeners = new Set<(name: string) => void>();
-
-/**
- * Fetch school name from the backend (best-effort).
- * Call after initializeApi(). Failure is silently ignored.
- */
-export async function fetchSchoolName(): Promise<void> {
-  try {
-    const res = await apiCall<{ status: string; data: { name: string } }>('/api/iot/school-name', {
-      headers: { Authorization: `Bearer ${DEVICE_API_KEY}` },
-    });
-    schoolNameState.value = res.data.name;
-    logger.info('School name loaded', { schoolName: schoolNameState.value });
-    for (const listener of schoolNameListeners) listener(schoolNameState.value);
-    schoolNameListeners.clear();
-  } catch {
-    logger.warn('Failed to fetch school name, continuing without it');
-  }
-}
-
-/**
- * Ensure API is initialized before making calls
- */
-async function ensureInitialized(): Promise<void> {
-  if (!isInitialized) {
-    await initializeApi();
-  }
-}
-
-/**
- * Handle network-level fetch errors and convert to user-friendly German messages
- * Extracted to reduce cognitive complexity in apiCall
- */
-function handleNetworkError(error: unknown, endpoint: string, startTime: number): never {
-  const errorObj = error instanceof Error ? error : new Error(String(error));
-  const responseTime = Date.now() - startTime;
-
-  logger.warn('Network error', {
-    endpoint,
-    responseTime,
-    errorName: errorObj.name,
-    errorMessage: errorObj.message,
-  });
-
-  reportNetworkStatus(responseTime, false);
-
-  // Map error types to German messages
-  if (errorObj.name === 'TypeError' && errorObj.message.includes('fetch')) {
-    throw new Error('Keine Netzwerkverbindung. Bitte WLAN prüfen.');
-  }
-  if (errorObj.name === 'AbortError') {
-    throw new Error('Zeitüberschreitung. Server antwortet nicht.');
-  }
-  if (errorObj.message.includes('NetworkError') || errorObj.message.includes('network')) {
-    throw new Error('Netzwerkfehler. Bitte Verbindung prüfen.');
-  }
-  throw new Error('Verbindungsfehler. Bitte Netzwerkverbindung prüfen.');
-}
-
-/**
- * Parse error response body and extract structured error data
- * Extracted to reduce cognitive complexity in apiCall
- */
-async function parseErrorResponse(
-  response: Response,
-  baseMessage: string
-): Promise<{ message: string; code?: string; details?: ApiErrorResponse['details'] }> {
-  try {
-    const errorData = (await response.json()) as ApiErrorResponse;
-    // Extract error detail from response - check message first, then error field
-    const errorDetail = errorData.message ?? (errorData as { error?: string }).error;
-    const message = errorDetail ? `${baseMessage}: ${errorDetail}` : baseMessage;
-    return { message, code: errorData.code, details: errorData.details };
-  } catch {
-    // JSON parsing failed, use base message
-    return { message: baseMessage };
-  }
-}
-
-/**
- * Generic API call function with error handling and response timing
- */
-async function apiCall<T>(endpoint: string, options: RequestInit = {}): Promise<T> {
-  await ensureInitialized();
-
-  const url = `${API_BASE_URL}${endpoint}`;
-  const startTime = Date.now();
-
-  let response: Response;
-  try {
-    response = await fetch(url, {
-      ...options,
-      headers: {
-        'Content-Type': 'application/json',
-        ...options.headers,
-      },
-    });
-  } catch (error) {
-    handleNetworkError(error, endpoint, startTime);
-  }
-
-  const responseTime = Date.now() - startTime;
-
-  if (!response.ok) {
-    const baseMessage = `API Error: ${response.status} - ${response.statusText}`;
-    const { message, code, details } = await parseErrorResponse(response, baseMessage);
-
-    logger.warn('API request failed', {
-      endpoint,
-      status: response.status,
-      responseTime,
-      error: message,
-      errorCode: code,
-      errorDetails: details,
-    });
-
-    throw new ApiError(message, response.status, code, details);
-  }
-
-  logger.debug('API request completed', {
-    endpoint,
-    status: response.status,
-    responseTime,
-    quality: responseTime < POOR_THRESHOLD_MS ? 'online' : 'poor',
-  });
-
-  reportNetworkStatus(responseTime, true);
-
-  return response.json() as Promise<T>;
-}
 
 /**
  * Teacher data structure from API
@@ -806,9 +174,7 @@ export const api = {
    */
   async getTeachers(): Promise<Teacher[]> {
     const response = await apiCall<TeacherResponse>('/api/iot/teachers', {
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-      },
+      headers: buildAuthHeaders(),
     });
 
     return response.data;
@@ -822,15 +188,12 @@ export const api = {
     try {
       logger.debug('Starting global PIN validation', {
         pin: pin.length + ' digits',
-        hasApiKey: !!DEVICE_API_KEY,
+        hasApiKey: hasDeviceApiKey(),
       });
 
       await apiCall('/api/iot/ping', {
         method: 'POST',
-        headers: {
-          Authorization: `Bearer ${DEVICE_API_KEY}`,
-          'X-Staff-PIN': pin,
-        },
+        headers: buildAuthHeaders(pin),
       });
 
       logger.info('Global PIN validation successful');
@@ -866,10 +229,7 @@ export const api = {
    */
   async getActivities(pin: string): Promise<ActivityResponse[]> {
     const response = await apiCall<ActivitiesResponse>('/api/iot/activities', {
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
     });
 
     return response.data;
@@ -881,7 +241,7 @@ export const api = {
    * Note: Returns plain text "OK", not JSON, so we don't use apiCall
    */
   async healthCheck(): Promise<void> {
-    const response = await fetch(`${API_BASE_URL}/health`, {
+    const response = await fetch(`${getApiBaseUrl()}/health`, {
       method: 'GET',
     });
 
@@ -906,10 +266,7 @@ export const api = {
     const endpoint = `/api/iot/rooms/available${queryPart}`;
 
     const response = await apiCall<RoomsResponse>(endpoint, {
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
     });
 
     return response.data;
@@ -928,10 +285,7 @@ export const api = {
       message?: string;
     }>('/api/iot/session/start', {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify(request),
     });
 
@@ -949,10 +303,7 @@ export const api = {
         data: CurrentSession | { device_id: number; is_active: false };
         message?: string;
       }>('/api/iot/session/current', {
-        headers: {
-          Authorization: `Bearer ${DEVICE_API_KEY}`,
-          'X-Staff-PIN': pin,
-        },
+        headers: buildAuthHeaders(pin),
       });
 
       // Check if we have an active session
@@ -979,10 +330,7 @@ export const api = {
   async endSession(pin: string): Promise<void> {
     await apiCall('/api/iot/session/end', {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
     });
   },
 
@@ -1006,10 +354,7 @@ export const api = {
       message?: string;
     }>(`/api/iot/session/${sessionId}/supervisors`, {
       method: 'PUT',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify({ supervisor_ids: supervisorIds }),
     });
 
@@ -1029,10 +374,7 @@ export const api = {
       data: Student[];
       message: string;
     }>(`/api/iot/students${queryParam}`, {
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
     });
 
     return response.data;
@@ -1049,10 +391,7 @@ export const api = {
         message: string;
         data: TagAssignmentCheck;
       }>(`/api/iot/rfid/${tagId}`, {
-        headers: {
-          Authorization: `Bearer ${DEVICE_API_KEY}`,
-          'X-Staff-PIN': pin,
-        },
+        headers: buildAuthHeaders(pin),
       });
 
       return response.data;
@@ -1084,10 +423,7 @@ export const api = {
       message?: string;
     }>(`/api/students/${studentId}/rfid`, {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify({
         rfid_tag: tagId,
       }),
@@ -1122,11 +458,7 @@ export const api = {
       message?: string;
     }>(`/api/iot/staff/${staffId}/rfid`, {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-        'X-Staff-ID': staffId.toString(),
-      },
+      headers: buildAuthHeaders(pin, staffId),
       body: JSON.stringify({
         rfid_tag: tagId,
       }),
@@ -1160,11 +492,7 @@ export const api = {
       message?: string;
     }>(`/api/iot/staff/${staffId}/rfid`, {
       method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-        'X-Staff-ID': staffId.toString(),
-      },
+      headers: buildAuthHeaders(pin, staffId),
     });
 
     return {
@@ -1194,10 +522,7 @@ export const api = {
       message?: string;
     }>(`/api/students/${studentId}/rfid`, {
       method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
     });
 
     return {
@@ -1227,10 +552,7 @@ export const api = {
       status: string;
     }>('/api/iot/checkin', {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify(scanData),
     });
 
@@ -1263,10 +585,7 @@ export const api = {
       status: string;
     }>('/api/iot/pickup-query', {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify(scanData),
     });
 
@@ -1280,10 +599,7 @@ export const api = {
   async updateSessionActivity(pin: string): Promise<void> {
     await apiCall('/api/iot/session/activity', {
       method: 'POST',
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-        'X-Staff-PIN': pin,
-      },
+      headers: buildAuthHeaders(pin),
       body: JSON.stringify({
         activity_type: 'rfid_scan', // Changed from 'student_scan' to 'rfid_scan'
         timestamp: new Date().toISOString(),
@@ -1314,10 +630,7 @@ export const api = {
 
       const response = await apiCall<AttendanceToggleResponse>('/api/iot/attendance/toggle', {
         method: 'POST',
-        headers: {
-          Authorization: `Bearer ${DEVICE_API_KEY}`,
-          'X-Staff-PIN': pin,
-        },
+        headers: buildAuthHeaders(pin),
         body: JSON.stringify(body),
       });
 
@@ -1339,10 +652,7 @@ export const api = {
     try {
       const response = await apiCall<DailyFeedbackResponse>('/api/iot/feedback', {
         method: 'POST',
-        headers: {
-          Authorization: `Bearer ${DEVICE_API_KEY}`,
-          'X-Staff-PIN': pin,
-        },
+        headers: buildAuthHeaders(pin),
         body: JSON.stringify(feedback),
       });
 
@@ -1360,9 +670,7 @@ export const api = {
    */
   async getDeviceConfig(): Promise<DeviceConfig> {
     const response = await apiCall<{ status: string; data: DeviceConfig }>('/api/iot/config', {
-      headers: {
-        Authorization: `Bearer ${DEVICE_API_KEY}`,
-      },
+      headers: buildAuthHeaders(),
     });
 
     return response.data;
